requirements.txt
django==3.0.8
djangorestframework==3.11.0
flake8==3.8.3
psycopg2==2.8.5
drf-yasg==1.17.1
Pillow==7.2.0
